refactor: remove obsolete chatd code and tests
