+++ coderd/x/chatd/chattool/mcpworkspace_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"net/http"
 	"strings"
-	"sync/atomic"
 	"testing"
 
 	"charm.land/fantasy"
@@ -27,18 +26,17 @@ func (f *fakeAgentConn) CallMCPTool(ctx context.Context, req workspacesdk.CallMC
 	return f.callMCPToolFunc(ctx, req)
 }
 
-func TestWorkspaceMCPTool_InvalidateOn404(t *testing.T) {
+func TestWorkspaceMCPTool_Errors(t *testing.T) {
 	t.Parallel()
 
-	t.Run("404ErrorInvalidatesCache", func(t *testing.T) {
+	t.Run("NotFound", func(t *testing.T) {
 		t.Parallel()
 
-		var invalidated atomic.Bool
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
+		tool := chattool.NewWorkspaceMCPTools(
+			[]workspacesdk.MCPToolInfo{{
 				Name:        "test__echo",
 				Description: "test tool",
-			},
+			}},
 			func(ctx context.Context) (workspacesdk.AgentConn, error) {
 				return &fakeAgentConn{
 					callMCPToolFunc: func(_ context.Context, _ workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
@@ -52,25 +50,21 @@ func TestWorkspaceMCPTool_InvalidateOn404(t *testing.T) {
 					},
 				}, nil
 			},
-			func() { invalidated.Store(true) },
-		)
+		)[0]
 
 		resp, err := tool.Run(context.Background(), fantasy.ToolCall{})
 		require.NoError(t, err)
 		assert.True(t, resp.IsError, "response should be an error")
-		assert.True(t, invalidated.Load(),
-			"invalidateCache should fire on 404")
 	})
 
-	t.Run("Non404DoesNotInvalidate", func(t *testing.T) {
+	t.Run("BadGateway", func(t *testing.T) {
 		t.Parallel()
 
-		var invalidated atomic.Bool
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
+		tool := chattool.NewWorkspaceMCPTools(
+			[]workspacesdk.MCPToolInfo{{
 				Name:        "test__echo",
 				Description: "test tool",
-			},
+			}},
 			func(ctx context.Context) (workspacesdk.AgentConn, error) {
 				return &fakeAgentConn{
 					callMCPToolFunc: func(_ context.Context, _ workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
@@ -83,25 +77,21 @@ func TestWorkspaceMCPTool_InvalidateOn404(t *testing.T) {
 					},
 				}, nil
 			},
-			func() { invalidated.Store(true) },
-		)
+		)[0]
 
 		resp, err := tool.Run(context.Background(), fantasy.ToolCall{})
 		require.NoError(t, err)
 		assert.True(t, resp.IsError)
-		assert.False(t, invalidated.Load(),
-			"invalidateCache should NOT fire on non-404 error")
 	})
 
-	t.Run("ToolLevelErrorNoInvalidation", func(t *testing.T) {
+	t.Run("ToolError", func(t *testing.T) {
 		t.Parallel()
 
-		var invalidated atomic.Bool
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
+		tool := chattool.NewWorkspaceMCPTools(
+			[]workspacesdk.MCPToolInfo{{
 				Name:        "test__echo",
 				Description: "test tool",
-			},
+			}},
 			func(ctx context.Context) (workspacesdk.AgentConn, error) {
 				return &fakeAgentConn{
 					callMCPToolFunc: func(_ context.Context, _ workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
@@ -114,41 +104,8 @@ func TestWorkspaceMCPTool_InvalidateOn404(t *testing.T) {
 					},
 				}, nil
 			},
-			func() { invalidated.Store(true) },
-		)
+		)[0]
 
-		resp, err := tool.Run(context.Background(), fantasy.ToolCall{})
-		require.NoError(t, err)
-		assert.True(t, resp.IsError)
-		assert.False(t, invalidated.Load(),
-			"invalidateCache should NOT fire on tool-level error (HTTP 200)")
-	})
-
-	t.Run("NilInvalidateCallbackSafe", func(t *testing.T) {
-		t.Parallel()
-
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
-				Name:        "test__echo",
-				Description: "test tool",
-			},
-			func(ctx context.Context) (workspacesdk.AgentConn, error) {
-				return &fakeAgentConn{
-					callMCPToolFunc: func(_ context.Context, _ workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
-						return workspacesdk.CallMCPToolResponse{}, codersdk.NewError(
-							http.StatusNotFound,
-							codersdk.Response{
-								Message: "MCP tool call failed.",
-								Detail:  `unknown MCP server: "test"`,
-							},
-						)
-					},
-				}, nil
-			},
-			nil,
-		)
-
-		// Should not panic.
 		resp, err := tool.Run(context.Background(), fantasy.ToolCall{})
 		require.NoError(t, err)
 		assert.True(t, resp.IsError)
@@ -162,13 +119,13 @@ func TestWorkspaceMCPTool_SanitizesModelNameKeepsRoutingName(t *testing.T) {
 		t.Parallel()
 
 		var gotToolName string
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
+		tool := chattool.NewWorkspaceMCPTools(
+			[]workspacesdk.MCPToolInfo{{
 				// "@" is outside the provider's allowed tool-name set; the
 				// model must never see it or the whole request is rejected.
 				Name:        "weather@home__get_forecast",
 				Description: "test tool",
-			},
+			}},
 			func(_ context.Context) (workspacesdk.AgentConn, error) {
 				return &fakeAgentConn{
 					callMCPToolFunc: func(_ context.Context, req workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
@@ -179,8 +136,7 @@ func TestWorkspaceMCPTool_SanitizesModelNameKeepsRoutingName(t *testing.T) {
 					},
 				}, nil
 			},
-			nil,
-		)
+		)[0]
 
 		// The model-facing name is sanitized to the provider-safe set.
 		assert.Equal(t, "weather_home__get_forecast", tool.Info().Name)
@@ -196,11 +152,11 @@ func TestWorkspaceMCPTool_SanitizesModelNameKeepsRoutingName(t *testing.T) {
 	t.Run("ValidNameUnchanged", func(t *testing.T) {
 		t.Parallel()
 
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
+		tool := chattool.NewWorkspaceMCPTools(
+			[]workspacesdk.MCPToolInfo{{
 				Name:        "github__create_issue",
 				Description: "test tool",
-			},
+			}},
 			func(_ context.Context) (workspacesdk.AgentConn, error) {
 				return &fakeAgentConn{
 					callMCPToolFunc: func(_ context.Context, _ workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
@@ -208,8 +164,7 @@ func TestWorkspaceMCPTool_SanitizesModelNameKeepsRoutingName(t *testing.T) {
 					},
 				}, nil
 			},
-			nil,
-		)
+		)[0]
 
 		// A name already within the allowed set is left untouched.
 		assert.Equal(t, "github__create_issue", tool.Info().Name)
@@ -221,11 +176,11 @@ func TestWorkspaceMCPTool_SanitizesModelNameKeepsRoutingName(t *testing.T) {
 		// A name longer than the provider limit is truncated. "srv__" plus a
 		// 64-char tool name exceeds the 64-char cap.
 		longName := "srv__" + strings.Repeat("a", 64)
-		tool := chattool.NewWorkspaceMCPTool(
-			workspacesdk.MCPToolInfo{
+		tool := chattool.NewWorkspaceMCPTools(
+			[]workspacesdk.MCPToolInfo{{
 				Name:        longName,
 				Description: "test tool",
-			},
+			}},
 			func(_ context.Context) (workspacesdk.AgentConn, error) {
 				return &fakeAgentConn{
 					callMCPToolFunc: func(_ context.Context, _ workspacesdk.CallMCPToolRequest) (workspacesdk.CallMCPToolResponse, error) {
@@ -233,8 +188,7 @@ func TestWorkspaceMCPTool_SanitizesModelNameKeepsRoutingName(t *testing.T) {
 					},
 				}, nil
 			},
-			nil,
-		)
+		)[0]
 
 		// The model-facing name is capped at the strictest provider limit.
 		assert.LessOrEqual(t, len(tool.Info().Name), 64)
@@ -261,7 +215,7 @@ func TestNewWorkspaceMCPTools_DisambiguatesCollidingNames(t *testing.T) {
 		{Name: "foo_bar__echo"},
 	}
 
-	tools := chattool.NewWorkspaceMCPTools(infos, getConn, nil)
+	tools := chattool.NewWorkspaceMCPTools(infos, getConn)
 	require.Len(t, tools, 2)
 
 	names := []string{tools[0].Info().Name, tools[1].Info().Name}
