+++ docs/admin/integrations/prometheus.md
@@ -250,7 +250,6 @@ The `coder_ai_gateway_cost_control_*` metrics are exported only by `coderd`.
 | `coderd_chatd_message_count`                                             | histogram | Number of messages in the prompt per LLM request.                                                                                                                                                                                                                                                                                                                                                                                                                                                          | `model` `provider`                                                                                    |
 | `coderd_chatd_prompt_size_bytes`                                         | histogram | Estimated byte size of the prompt per LLM request.                                                                                                                                                                                                                                                                                                                                                                                                                                                         | `model` `provider`                                                                                    |
 | `coderd_chatd_steps_total`                                               | counter   | Total agentic loop steps across all chats.                                                                                                                                                                                                                                                                                                                                                                                                                                                                 | `model` `provider`                                                                                    |
-| `coderd_chatd_stream_buffer_dropped_total`                               | counter   | Number of chat stream buffer events dropped due to the per-chat buffer cap.                                                                                                                                                                                                                                                                                                                                                                                                                                |                                                                                                       |
 | `coderd_chatd_stream_retries_total`                                      | counter   | Total LLM stream retries.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  | `kind` `model` `provider`                                                                             |
 | `coderd_chatd_tool_errors_total`                                         | counter   | Total tool calls that returned an error result.                                                                                                                                                                                                                                                                                                                                                                                                                                                            | `model` `provider` `tool_name`                                                                        |
 | `coderd_chatd_tool_result_size_bytes`                                    | histogram | Size in bytes of each tool execution result.                                                                                                                                                                                                                                                                                                                                                                                                                                                               | `model` `provider` `tool_name`                                                                        |
